project/baseline/references/requirements_distillation.txt
# Enhanced Requirements for Task Distillation
# Based on baseline requirements with additional distillation dependencies

# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Integration (for teacher models)
openai>=1.0.0
anthropic>=0.7.0
requests>=2.28.0

# Evaluation and Metrics
scikit-learn>=1.3.0
scipy>=1.11.0
nltk>=3.8.1
rouge-score>=0.1.2
sacrebleu>=2.3.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
wandb>=0.15.0  # For experiment tracking
tensorboard>=2.14.0

# Quantization and Optimization
bitsandbytes>=0.41.0
optimum>=1.14.0

# Audio Processing (existing from baseline)
funasr>=1.0.0
soundfile>=0.12.1

# API Server (existing from baseline)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Monitoring (existing from baseline)
psutil>=5.9.6
GPUtil>=1.4.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
